feat: move game logic from component to hook
