+++ src/components/GameBoard/GameBoard.tsx
@@ -1,33 +1,10 @@
-import { useEffect } from 'react';
 import styled from 'styled-components';
 import { Cell } from './Cell';
-import {
-  checkVictory,
-  createInitGrid,
-  defineNextStatusCell,
-  fillGrid,
-  getCountFlags,
-  openArea,
-  openBombsAfterLost,
-  openCellsAfterWin,
-} from './utils';
 
-import { useAppSelector, useAppDispatch } from '../../hooks/redux';
-import {
-  changeCells,
-  changePhase,
-  selectState,
-  updateOneCell,
-  changeBombsLeft,
-  changeTimeleft,
-} from '../../store/reducers/gameSlice';
-
-import { ICell, IWinner } from '../../types/types';
+import { ICell } from '../../types/types';
 import { Frame } from '../Frame';
 import { GamePanel } from './GamePanel';
-import { useTimer } from '../../hooks/timer';
-import { addWinner } from '../../store/reducers/winnersSlice';
-import { localStorageKey } from '../../const/const';
+import { useGame } from '../../hooks/game/useGame';
 
 interface IGridTemplateProps {
   col: number;
@@ -52,149 +29,15 @@ const GridTemplate = styled.div<IGridTemplateProps>`
 `;
 
 function GameBoard() {
-  const { startTimer, stopTimer, timeLeft } = useTimer();
-  const dispatch = useAppDispatch();
-  const { phase, boardParams, cells, bombsLeft } = useAppSelector(selectState);
+  const { cells, boardParams, clickLeftButton, clickRightButton } = useGame();
 
-  const prepareNewGame = () => {
-    const initGrid = createInitGrid({ ...boardParams });
-    dispatch(changeCells(initGrid));
+  const clickCellHandler = (cell: ICell) => {
+    clickLeftButton(cell);
   };
 
-  const startNewGame = (cell: ICell) => {
-    if (cell.status === 'flag-icon') return;
-
-    stopTimer();
-    startTimer();
-
-    const filledGrid: ICell[][] = fillGrid({
-      source: cells,
-      bombs: boardParams.bombs,
-      currentCellPosition: cell.position,
-    });
-    dispatch(changePhase('play'));
-    const { i, j } = cell.position;
-    if (filledGrid[i][j].content === 0) {
-      dispatch(changeCells(openArea(filledGrid, filledGrid[i][j])));
-    } else {
-      dispatch(changeCells(filledGrid));
-      dispatch(updateOneCell({ ...filledGrid[i][j], status: 'around-bombs' }));
-    }
+  const clickContextCellHandler = (cell: ICell) => {
+    clickRightButton(cell);
   };
-
-  const endGameWithVictory = () => {
-    const winner: IWinner = {
-      id: Math.random(),
-      timeLeft,
-      boardParams,
-    };
-
-    dispatch(changePhase('win'));
-    dispatch(
-      addWinner(winner),
-    );
-  };
-
-  const makeMove = (cell: ICell) => {
-    // do nothing if flag
-    if (cell.status === 'flag-icon') return;
-
-    let isVictory = false;
-
-    // game over if click to bomb
-    if (cell.content === -1) {
-      dispatch(changePhase('lost'));
-      dispatch(updateOneCell({ ...cell, status: 'bomb-boom' }));
-    }
-
-    // open area if click in empty cell
-    if (cell.content === 0) {
-      const newCells = openArea(cells, cell);
-      dispatch(changeCells(newCells));
-      isVictory = checkVictory(newCells, bombsLeft, null);
-    }
-
-    // open one cell if click in cell with digit
-    if (cell.content > 0) {
-      dispatch(updateOneCell({ ...cell, status: 'around-bombs' }));
-      isVictory = checkVictory(cells, bombsLeft, {
-        ...cell,
-        status: 'around-bombs',
-      });
-    }
-
-    if (isVictory) {
-      endGameWithVictory();
-    }
-  };
-
-  const clickCellHandle = (cell: ICell) => {
-    if (phase === 'new' || phase === 'change-lvl') {
-      startNewGame(cell);
-    }
-    if (phase === 'play' && cell.status === 'closed') {
-      makeMove(cell);
-    }
-  };
-
-  const clickContextCellHandle = (cell: ICell) => {
-    if (phase !== 'lost' && phase !== 'win') {
-      const status = defineNextStatusCell(cell);
-      let isVictory = false;
-
-      if (status !== null) {
-        dispatch(updateOneCell({ ...cell, status }));
-      }
-
-      let countFlags = getCountFlags(cells);
-      if (status === 'quest-icon') countFlags -= 1;
-      if (status === 'flag-icon') countFlags += 1;
-      dispatch(changeBombsLeft(boardParams.bombs - countFlags));
-
-      isVictory = checkVictory(
-        cells,
-        boardParams.bombs - countFlags,
-        status === null ? null : { ...cell, status: status || '' },
-      );
-      if (isVictory) {
-        endGameWithVictory();
-      }
-    }
-  };
-
-  const restartGame = () => {
-    stopTimer();
-    prepareNewGame();
-    dispatch(changePhase('new'));
-    dispatch(changeBombsLeft(boardParams.bombs));
-  };
-
-  useEffect(() => {
-    if (phase === 'new' || phase === 'change-lvl') {
-      restartGame();
-      stopTimer();
-      dispatch(changeTimeleft(0));
-    }
-    if (phase === 'win') {
-      const newCells = openCellsAfterWin(cells);
-      dispatch(changeCells(newCells));
-      stopTimer();
-    }
-    if (phase === 'lost') {
-      const newCells = openBombsAfterLost(cells);
-      dispatch(changeCells(newCells));
-      stopTimer();
-    }
-  }, [phase]);
-
-  useEffect(() => {
-    if (timeLeft > 999) {
-      dispatch(changePhase('lost'));
-      return;
-    }
-
-    dispatch(changeTimeleft(timeLeft));
-  }, [timeLeft]);
 
   return (
     <Frame variant="outside">
@@ -207,8 +50,8 @@ function GameBoard() {
                 <Cell
                   key={cell.id}
                   cell={cell}
-                  clickCellHandle={clickCellHandle}
-                  clickContextCellHandle={clickContextCellHandle}
+                  clickCellHandle={clickCellHandler}
+                  clickContextCellHandle={clickContextCellHandler}
                 />
               )),
             )}
