+++ src/components/GameMenu/GameMenu.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import styled from 'styled-components';
 import { levels, menuButtons } from '../../const/const';
 import { TLevelType } from '../../types/types';
-import { getBoardParamsByLevelType } from '../GameBoard/utils';
+import { getBoardParamsByLevelType } from '../../hooks/game/utils';
 
 import { useAppDispatch, useAppSelector } from '../../hooks/redux';
 import { changeLevel, changePhase } from '../../store/reducers/gameSlice';
