+++ src/hooks/game/useTimer.ts
