+++ src/hooks/game/utils.ts
